create a duchess-reflect crate
This crate contains the core reflection logic
that will be shared between build.rs script
support and procedural macros.
